requirements.txt
python-telegram-bot
lyricsgenius